requirements.txt
Flask
python-dotenv
flake8
requests